Remove oslo-incubator code
Because of the necessity of importing code
directly from Nova and the fact that Nova includes and imports
its own incubated code, using our own copies of incubated code
can conflict. As a practical example, the oslo logging module
sets configuration variables (using oslo.config) that conflict
with multiple imports of that same code.

While it might be possible to catch these exceptions and handle
them within the Oslo modules, it's relatively unnecessary given
that we must already depend upon and import code from Nova.

This might prove problematic later should we need to bring in
oslo modules that are not imported in Nova, but that's a bridge
we can cross if and when that becomes more than an academic exercise.

Change-Id: I5a612221e2a77768fcab45329726488b55fd5f8e