Install nova (master/git) in testing environment
As we're a plugin, we require several imports from nova.
Thus, it is important that we have those modules available in
our testing environment.

Unfortunately, it seems that bringing this in via the requirements
files is unreliable so we instead include this directly via the tox.ini
where we can specify the '-e' parameter.

Change-Id: Ia5aa9d91fe5c263f5c8d492f9335c6e9adb191eb