OpenContrail support for nova-docker
Following commits are bundeled:

Make the vif_driver configurable -Pedro Marques

Initial support for OpenContrail -Pedro Marques

Revert "Make the vif_driver configurable" -Martin Loschwitz

This reverts commit 4460711abd3e87317e4c5b24eee26aa96cdd8990.

Split "plug" function into a "plug" and an "attach" function -Martin
Loschwitz

Recent versions of docker expect two functions, one named "plug" and the
other named "attach". This commit changes the code so that both functions
exist and do the right thing to make Docker happy. Without the patch, the
correct association of interfaces and network namespaces is broken (i.e.
the driver will try to add an interface to a namespace that does nut exist
yet).

Adapt this driver to the new i18n framework in Nova -Martin Loschwitz

Adapt to latest changes in OpenContrail -Martin Loschwitz

Don't make the driver crash nova-compute if veth if already exists
-Martin Loschwitz

In its current form, the opencontrail driver for docker will try to create
a veth interface even if it's already there. This will result in an error
that will make nova-compute crash. In fact, nova-compute is unrestartable
with the current driver. This patch adds a check to see if the veth if is
already there and does not try to add it a second time if it already is.

removed thrift call, using ContrailVRouterApi -Ted Ghose

Change-Id: I1d032143a4c346d3aa7169017301201c278098d3