Adds ability to use a remote registry
Adds the ability to the docker driver to connect to a remote docker
registry instead of assuming the registry is always running locally.
This allows different docker compute nodes to connect to the same docker
registry without each node having to run its own.

Adds a new config option to the docker group called
registry_default_host where you can specify the hostname or IP address
of the registry host. The option defaults to my_ip from nova.netconf
which preserves the existing behavior of using a locally running
registry.

Change-Id: I3b8a4fbca1be969da803217794f094f8cd83b082