Report correct vcpus and memory in instance info
Fetches memory and cpushares from Docker and returns
them upon requests for instance info as max_mem, mem,
and num_cpu.

The used memory is reported as the maximum memory as
we cannot yet fetch statistics from the memory cgroup.
A FIXME is commented to improve upon this and report
the correct amount of memory utilized by the cgroup
processes.

Furthermore, a comment notes further improvements
possible for reporting cpu_time.

Closes-Bug: 1249108

Change-Id: Idce00dc094c2ea64450eaf74c94a8912a5961e0e