Implement soft_delete and restore
Adds the soft_delete and restore features.
This also refactors a bit how we start/stop/destroy containers.

Adds a test for testing soft deletes and restores.

Also fixes a KeyError in the mock_client where 'id' was
used as a key where the correct key would have been 'Id'.

Closes-Bug: #1321819

Change-Id: Icae0d521bb471b5fd43a4adb8d51d3f906766cfe