Adds support for rebooting docker containers
Without this patch, whenever a container is rebooted in OpenStack it
will loose its IP address.

The reason is that the container PID will change, but the namespace
is not updated to the new PID.
Further operations in that namespace will not be executed, and the
container effectivelly becomes isolated.

The fix proposed is to destroy all networking and create it again.

Change-Id: I583926e43c832605c77852205b4e2df8ce9746ed