Split plug_vifs into plug_vifs and attach_vifs
Driver.plug_vifs is sometimes called when the container is not
running, such as called by _init_instance in ComputeManager.

In this patch, plug_vifs is only responsible for creating bridges
and preparing the peer devices, which does not require the container
running, attach_vifs is responsible for preparing the netns and
attaching the peer device to container.

When starting a container, attach_vifs is called after plug_vifs.

Change-Id: Icc5e2cebde22c10a318379d41bb4c50085b07162
Closes-bug: #1341423