Disable tempest
We've discovered that successfully running
devstack is enough to get tempest to start gating.

However, a yet-unknown bug is causing the gate VM
to crash, or the VM to become inaccessible.

Until we finish debugging this, we will disable
tempest to allow fixes to flow in and patches to be merged.

Change-Id: Ief05ae1220488117c17a5147321b631c3ca1a4e2