Tempest test configuration for CI check job.
This patch removes the post_test_hook that would launch
an instance and verify it was running. Instead of the post_test_hook,
this patch enables tempest tests which will test a superset of features.
The gate_hook.sh now sets up a regular expression to filter out the
tempest tests that are currently broken or not applicable to containers.

Co-Authored-By: Chris Alfonso <calfonso@redhat.com>
Change-Id: Ia6983d03a67d2ccf268040d9410456f4f2179b76