Ensure the bridge is correct and use high range MACs.
This patch adds a call to linux_net.ensure_bridge() to make
sure that the bridge exists and is correct.  It also creates
mac addresses for the tap devices that start with FE.  It turns
linux bridges use the lowest MAC available as the interface
address.  When we run tempest tests we create so many instances
that the mac can change rapidly causing a loss of networking.

Closes-Bug: #1339679

Co-Authored-by: Chris Alfonso <calfonso@redhat.com>

Change-Id: I5ae5f27e07a4b9de98a5a80f89ac2d7e8c0603f7