Support timeout on power_off (and stop_container)
Nova now has a power_off timeout. We will now support
this and pass Nova's timeout into stop_container.

The timeout in Nova is now 60 seconds so this change will
effectively increase the container_stop timeout from 5 seconds
to 60 seconds. That will be the worst-case performance in the
event that the process(es) in the container do not respond
to a SIGTERM.

Change-Id: Idffd3ad96989085c97647cb731ae0448cb47a05c