Inject SSH keys into the docker container
Create a directory for .ssh, create a authorized_keys file with
the correct permissions, then bind mount the .ssh directory when
we start the container. This enables anyone to create a ssh
key-pair in nova specify it in the nova boot command. This is
an alternative mechanism modeled after libvirt's inject_key
flag for environments that don't have metadata service (and
since we don't have config drive support).

DocImpact

Change-Id: I71f87e9da62eee73a641d20be6bd6cafce4c0f3d