Pass launch metadata as enviroment variables
The launch metadata is provided with 'nova boot' and
serves effectively to pass varibles into an instance. This
runtime key-value pairing seems to map cleanly to environment
variables as used by Docker. Use of environment variables
is highly desired amongst nova-docker users and this seems to
be the cleanest solution to deliver this feature.

Change-Id: I0292bf065191c6a188e39a7d2f2c0a5057ec67f5