Add a new script for copying logs
post_test_hook is not a good place to collect logs since it
does not get executed when jobs fail. Let's try executing
a script in our gate_hook to copy files over. Also use
a find/exec to copy the files since some files may be gzipped

We also need to copy over the log files for containers that
remain when we finish testing.

Change-Id: I37b50eff7b086b3a29849a292a2171744460da06