Unpause the container before stopping
If we get an error from Docker API when we issue a stop,
then we should call pause followed by a stop. We were doing
this in one scenario. This commit ensures that all docker.stop(s)
are called from one place with the try/except to ensure
pause gets called. This hopefully reduces failures related to
unmounting in aufs / devicemapper etc.

Change-Id: I95fa7a5af58f0587a8b72f4652a3e5636bff39a9