Give some time for docker.stop to work
REST API "/stop?t=0" ends up with
"failed to exit within 0 seconds of SIGTERM - using the force"
in the docker logs. So let's make sure we give a
bit of time for the docker daemon to do its job.
Note, this SIGTERM ends up with
"Driver aufs failed to remove root filesystem"
as well in the docker logs. So the timeout is very
important.

Closes-Bug: #1403439
Change-Id: If9bb60cc55433f8b2ec8e43e3829d273461a9015