use instance.flavor where appropriate
This changes replaces two bits of code with references to
instance.flavor.

In _get_memory_limit_bytes, this replaces a call to
system_meta.get('instance_type_memory_mb', 0) with
instance.flavor.memory_mb, which seems to be the intent there (i.e., get
the memory limit for the flavor).

In _get_cpu_shares, this replaces a call to flavors.extract_flavor
because, as the comment in nova/compute/flavors.py says:

> NOTE(danms): This method is deprecated, do not use it!
> Use instance.{old_,new_,}flavor instead, as instances no longer
> have flavor information in system_metadata.

Change-Id: I3592ba77d7182364f42592021484b2fd468de5d5