Fix tests on stable/juno
In stable/juno the tox.ini file is pulling nova from master which is
kilo-based. This patch set updates the test requirements to pull nova
from stable/juno.

This patch set also fixes pep8 errors that were fixed in kilo under
8bcecf6faa6a55bf081f5a4a4dad817db768094b and updates the tests so
that they are invoked with a lockpath in the same manner as in
Ib8f7b5b36568195da611892fdcc94354d03a14e7

Change-Id: Ifc149e14f46c243edd8ffeff860a056d379957dc