Fix docker daemon 1.7 startup option
Docker 1.7 makes udev sync a requirement for devicemapper
(https://github.com/docker/docker/pull/11412/commits), which
causes that docker daemon fails to start if udev sync is not
supported. This patch sets the option 'dm.override_udev_sync_check',
so that docker daemon can start with devicemapper and simply warn
with the error.

Change-Id: I8e1be63e5ec4c73a0a5e385d0809ac51d3ebee30