Setup for translation
To start translation, we need to initially import the
translation file - and place it at the proper place so that
the usual CI scripts can handle it.

The proper place is for all python projects
$PROJECT/locale/$PROJECT.pot - see setup.cfg.

Note that we should be using our own i18n.py
module per best practices:
http://docs.openstack.org/developer/oslo.i18n/usage.html#creating-an-integration-module

Further imports will be done by the OpenStack Proposal bot.

Closes-Bug: #1450385
Change-Id: Ie72fb14aba586f77acb2d5d9ec4821f7c95eaa7a